Add Window resize, logs, and pseudo vsync.
Wanted to add window resize to make sure the letterboxing was sound,
which it is. While debugging discovered that logging with printf was
being delayed 'till after the program quit, so switched to using
SDL_Log. Added a quick color swap effect to test the blitting and after
getting tearing, decided to add a pseudo vsync with performance counters
and SDL_Delay, a log is there to ensure the difference is never greater
than 2 milliseconds.